drm: add crtc background color property
Some display controllers can be programmed to present non-black colors
for pixels not covered by any plane (or pixels covered by the
transparent regions of higher planes).  Compositors that want a UI with
a solid color background can potentially save memory bandwidth by
setting the CRTC background property and using smaller planes to display
the rest of the content.

To avoid confusion between different ways of encoding RGB data, we
define a standard 64-bit format that should be used for this property's
value.  Helper functions and macros are provided to generate and dissect
values in this standard format with varying component precision values.

Signed-off-by: Raphael Gallais-Pou <raphael.gallais-pou@foss.st.com>
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>
Change-Id: Iebc91ed5707052fa98db1b052a762ae7edefc1fa
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/211001
Reviewed-by: Philippe CORNU <philippe.cornu@foss.st.com>
Reviewed-by: CIBUILD <MDG-smet-aci-builds@list.st.com>