spi: stm32: finalize message either on dma callback or EOT
Depending on the usage, it is necessary to perform the finalize
message operation either upon receiving the EOT interruption,
eiher upon receiving the DMA callback. Indeed, when relying
on DMA, even if the SPI EOT IT has been received, it is
necessary to wait for the end of the DMA RX transaction before
accessing to the data.

Signed-off-by: Alain Volmat <alain.volmat@foss.st.com>
Change-Id: Iab963e66c9ad27e63b6b642c9e2779e815a99063
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/198439
Reviewed-by: MPUOSTL
Reviewed-by: Fabrice GASNIER <fabrice.gasnier@foss.st.com>
Reviewed-by: Amelie DELAUNAY <amelie.delaunay@foss.st.com>