serial: stm32: avoid clearing DMAT bit during transfer
It's rather advised to rely on DMA pause / resume instead of
clearing/setting DMA request enable bit for the same purpose. Some DMA
request/acknowledge race may encountered by doing so. We prefer to use
dmaengine_pause and resume instead to pause a dma transfer when it is
necessary.

It is also safer to close DMA before reset DMAT in stm32_usart_shutdown.

Signed-off-by: Valentin Caron <valentin.caron@foss.st.com>
Change-Id: I5fc1dd455d2923a80fc41e5bad3a3de705496c3c
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/206112
Reviewed-by: Fabrice GASNIER <fabrice.gasnier@foss.st.com>
Reviewed-by: MPUOSTL
Reviewed-by: Amelie DELAUNAY <amelie.delaunay@foss.st.com>