net: ethernet: stmmac: stm32: change warning to debug message
There is a new parameter in syscon in ethernet DT since stm32mp13
to manage several ethernet intances.
In STM32MP15, there is a message "Warning sysconfig register mask not set"
when this new parameter is not set.
To avoid too many messages,replace dev_info with pr_debug to print this message.

Signed-off-by: Christophe Roullier <christophe.roullier@st.com>
Change-Id: Ie14cfa1d74dd512e84324e245708234e38cb441f
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/211916
Reviewed-by: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Alexandre TORGUE <alexandre.torgue@foss.st.com>