media: stm32: dcmi: create a dma scatterlist based on max_seg_size value
Prior to submitting a transfer to the DMA, the client should first check
the capabilities of the DMA in term of maximum of each segment. This is
given by the dma_get_max_seg_size API. Based on that, if the transfer
is larger than what can be handled by the DMA, we split the transfer into
several scatterlist elements.

Change-Id: I474d2b8ab1b4737aaa3e32e4f2d9c40c08192d60
Signed-off-by: Alain Volmat <alain.volmat@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/221693
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Amelie DELAUNAY <amelie.delaunay@foss.st.com>
Reviewed-by: Philippe CORNU <philippe.cornu@foss.st.com>