media: stm32-dcmipp: add compose / crop support in byteproc
Add the compose and crop support within byteproc subdev.  In the HW
the decimation is done prior to the cropping hence configuration is
possible via a compose on the sink pad and a crop on the src pad
(aka on top of the compose result).

Apply behavior expressed within the V4L2 documentation [1] in term of
format propagation from the sink to the src pad and ordering of
configuration.  Since the byteproc cannot modify the format we can
simplify a bit the code without having to store the src pad format.

Decimation (via compose) can be up to /4 horizontally and / 2 vertically,
however depending on the input format not everything is possible since
the byteproc is performing processing at the byte level (not pixel), so
we should take care avoiding breaking pixel data.

[1] https://www.kernel.org/doc/html/latest/userspace-api/media/v4l/dev-subdev.html

Signed-off-by: Alain Volmat <alain.volmat@foss.st.com>
Change-Id: I0c70d8e725fcae08e2b6c3a2bd5d7ca384a5de8b
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/244370
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Hugues FRUCHET <hugues.fruchet@st.com>
Reviewed-by: CIBUILD <MDG-smet-aci-builds@list.st.com>
Tested-by: Hugues FRUCHET <hugues.fruchet@st.com>