media: stm32-dcmipp: vsync / frame_end handling
Rework of the buffer handling within the byte capture device so that
DCMIPP continuous mode can be used instead of the snapshot mode.

This change enables DCMIPP framerate control feature while improvingcapture
framerate performance by relaxing constraint on frame done processing
execution time.

Main modification is located in the split of handling between the VSYNC
signal and the FRAME END signal.  VSYNC signal indicates the moment when
shadow registers are applied as settings and thus the earliest moment
when settings for the next frame can be written (into the shadow
registers).  FRAME END signal indicates that DCMIPP has finished to write
into the buffer and that this buffer can be given back to the user.

The driver now handles 2 buffers (active buffer, aka the one
into which the DCMIPP is currently writing into, and the next buffer,
aka the one currently written into the shadow register and which will be
applied at the next VSYNC) in addition to the list of queued buffer.

Add two new counters:
   to track underrun: user is not queuing fast enough buffers so that
                      captured frames are lost.
   to track getting a frame end while active is NULL

Change-Id: Ibe6e2feb599f0479d9ce3665502bb06cd3158ed0
Signed-off-by: Alain Volmat <alain.volmat@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/244125
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Hugues FRUCHET <hugues.fruchet@st.com>
Reviewed-by: CIBUILD <MDG-smet-aci-builds@list.st.com>
Tested-by: Hugues FRUCHET <hugues.fruchet@st.com>