usb: typec: ucsi: stm32g0: don't disabke wakeup by default
When the "wakeup-source" flag is found, I2C core marks the device as
wakeup capable and enables the wakeup by default.
Up to now, choice was being made to disable the wakeup by default. Goal was
to simplify the user application, to be able to reach lowest possible
power state by default (standby) on stm32mp135f-dk board.
User had to specifically re-enable the type-c wakeup if needed.
Since the PWR IRQ and wakeup capability is now implemented by the secure
and non-secure side, the standby mode can be reached by default.
So, remove earlier limitation, which was hacky behing "ucsi" and "psy"
frameworks.

Signed-off-by: Fabrice Gasnier <fabrice.gasnier@foss.st.com>
Change-Id: If2593b0d440c97d1e0cbe9da435f17b6e63da547
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/265107
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Alexandre TORGUE <alexandre.torgue@foss.st.com>